+++ src/main/java/lotto/controller/LottoGame.java
@@ -5,6 +5,7 @@ import java.util.Set;
 
 import lotto.domain.*;
 import lotto.view.InputView;
+import lotto.view.InputViewExceptionHandler;
 import lotto.view.OutputView;
 
 public class LottoGame {
@@ -12,32 +13,32 @@ public class LottoGame {
     public void play() {
         Money purchaseAmount = inputPurchaseAmount();
         Lottos lottosManual = purchaseLottosManually(purchaseAmount);
-        Lottos lottosAutomatic = Player.purchaseLottosAutomatically(purchaseAmount, lottosManual.size());
+        Lottos lottosAutomatic = LottoPlayer.purchaseLottosAutomatically(purchaseAmount, lottosManual.size());
         OutputView.printLottos(lottosManual, lottosAutomatic);
         Result result = produceResult(purchaseAmount, lottosManual, lottosAutomatic);
         OutputView.printResult(result);
     }
 
     private Money inputPurchaseAmount() {
-        int input = LottoLogicExecutor.executeOrRepeatWithException(InputView::inputPurchaseAmount);
+        int input = InputViewExceptionHandler.handle(InputView::inputPurchaseAmount);
         return NonPlayerCharacter.translateToMoney(input);
     }
 
     private Lottos purchaseLottosManually(Money purchaseAmount) {
-        int numberToBuyLottoManually = LottoLogicExecutor.executeOrRepeatWithException(() -> InputView.inputNumberToBuyManually(purchaseAmount.toLottosSize()));
-        List<Set<Integer>> numbersBasket = LottoLogicExecutor.executeOrRepeatWithException(() -> InputView.inputnumbersBasketManually(numberToBuyLottoManually));
+        int numberToBuyLottoManually = InputViewExceptionHandler.handle(() -> InputView.inputNumberToBuyManually(purchaseAmount.toLottosSize()));
+        List<Set<Integer>> numbersBasket = InputViewExceptionHandler.handle(() -> InputView.inputnumbersBasketManually(numberToBuyLottoManually));
         List<Set<LottoNumber>> lottoNumbersBasket = NonPlayerCharacter.translateToLottoNumbersBasket(numbersBasket);
-        return Player.purchaseLottosManually(lottoNumbersBasket);
+        return LottoPlayer.purchaseLottosManually(lottoNumbersBasket);
     }
 
     private Result produceResult(Money purchaseAmount, Lottos lottosManual, Lottos lottosAutomatic) {
         Lotto winningLotto = declareWinningLotto();
-        int bonusNumber = LottoLogicExecutor.executeOrRepeatWithException(InputView::inputBonusNumber);
+        int bonusNumber = InputViewExceptionHandler.handle(InputView::inputBonusNumber);
         return NonPlayerCharacter.judgeResult(lottosManual, lottosAutomatic, winningLotto, bonusNumber, purchaseAmount);
     }
 
     private Lotto declareWinningLotto() {
-        List<Integer> winningNumbers = LottoLogicExecutor.executeOrRepeatWithException(InputView::inputLastWeekWinningNumbers);
+        List<Integer> winningNumbers = InputViewExceptionHandler.handle(InputView::inputLastWeekWinningNumbers);
         return NonPlayerCharacter.declareWinningLotto(winningNumbers);
     }
 }
