--- src/main/java/lotto/domain/LottoLogic.java
@@ -1,5 +1,0 @@
-package lotto.domain;
-
-public interface LottoLogic<V> {
-    V work();
-}
