+++ src/main/java/lotto/view/InputStrategy.java
@@ -1,0 +1,5 @@
+package lotto.view;
+
+public interface InputStrategy<V> {
+    V input();
+}
